New upstream version 202303.3

+++ rdkit/ML/AnalyzeComposite.py
@@ -27,7 +27,7 @@ Usage:  AnalyzeComposite [optional args] <models>
       -v: be verbose whilst screening
 """
 
-
+import pickle
 import sys
 
 import numpy
@@ -35,9 +35,7 @@ import numpy
 from rdkit.Dbase.DbConnection import DbConnect
 from rdkit.ML import ScreenComposite
 from rdkit.ML.Data import Stats
-from rdkit.ML.DecTree import TreeUtils, Tree
-import pickle
-
+from rdkit.ML.DecTree import Tree, TreeUtils
 
 __VERSION_STRING = "2.2.0"
 
@@ -65,11 +63,11 @@ def ProcessIt(composites, nToConsider=3, verbose=0):
           levels = TreeUtils.CollectLabelLevels(model, {}, 0, nToConsider)
           TreeUtils.CollectDescriptorNames(model, descNames, 0, nToConsider)
           for descId in levels.keys():
-            v = res.get(descId, numpy.zeros(nToConsider, numpy.float))
+            v = res.get(descId, numpy.zeros(nToConsider, float))
             v[levels[descId]] += 1. / nModels
             res[descId] = v
       for k in res:
-        v = globalRes.get(k, numpy.zeros(nToConsider, numpy.float))
+        v = globalRes.get(k, numpy.zeros(nToConsider, float))
         v += res[k] / nComposites
         globalRes[k] = v
       if verbose > 0:
@@ -113,12 +111,12 @@ def ErrorStats(conn, where, enrich=1):
   if not nPts:
     sys.stderr.write('no runs found\n')
     return None
-  overall = numpy.zeros(nPts, numpy.float)
-  overallEnrich = numpy.zeros(nPts, numpy.float)
+  overall = numpy.zeros(nPts, float)
+  overallEnrich = numpy.zeros(nPts, float)
   oCorConf = 0.0
   oInCorConf = 0.0
-  holdout = numpy.zeros(nPts, numpy.float)
-  holdoutEnrich = numpy.zeros(nPts, numpy.float)
+  holdout = numpy.zeros(nPts, float)
+  holdoutEnrich = numpy.zeros(nPts, float)
   hCorConf = 0.0
   hInCorConf = 0.0
   overallMatrix = None
@@ -276,8 +274,10 @@ def Usage():
 if __name__ == "__main__":
   import getopt
   try:
-    args, extras = getopt.getopt(sys.argv[1:], 'n:d:N:vX', ('skip',
-                                                            'enrich=', ))
+    args, extras = getopt.getopt(sys.argv[1:], 'n:d:N:vX', (
+      'skip',
+      'enrich=',
+    ))
   except Exception:
     Usage()
